@@ -1,4 +1,4 @@
-class AddSlugToTeam < ActiveRecord::Migration
+class AddSlugToTeam < ActiveRecord::Migration[5.2]
   def change
     add_column :teams, :slug, :string, default: ""
   end
