@@ -1,4 +1,4 @@
-class AddLeagueToGames < ActiveRecord::Migration
+class AddLeagueToGames < ActiveRecord::Migration[5.2]
   def change
     add_column :games, :league, :string
   end
