@@ -1,4 +1,4 @@
-class AddScheduleUrlToTeam < ActiveRecord::Migration
+class AddScheduleUrlToTeam < ActiveRecord::Migration[5.2]
   def change
     add_column :teams, :schedule_url, :string
   end
