@@ -1,4 +1,4 @@
-class AddTimezoneToTeam < ActiveRecord::Migration
+class AddTimezoneToTeam < ActiveRecord::Migration[5.2]
   def change
     add_column :teams, :timezone, :string
   end
