@@ -1,4 +1,4 @@
-class AddUniquenessConstraints < ActiveRecord::Migration
+class AddUniquenessConstraints < ActiveRecord::Migration[5.2]
   def change
     change_column :teams, :name, :string, unique: true
   end
