@@ -1,4 +1,4 @@
-class AddIndexToNicknameName < ActiveRecord::Migration
+class AddIndexToNicknameName < ActiveRecord::Migration[5.2]
   def change
     add_index :nicknames, :name
   end
