@@ -1,4 +1,4 @@
-class NicknamesController < ApplicationController
+class NicknamesController < AdminController
   before_action :set_nickname, only: [:show, :edit, :update, :destroy]
 
   # GET /nicknames
@@ -62,13 +62,14 @@ class NicknamesController < ApplicationController
   end
 
   private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_nickname
-      @nickname = Nickname.find(params[:id])
-    end
 
-    # Never trust parameters from the scary internet, only allow the white list through.
-    def nickname_params
-      params.require(:nickname).permit(:name, :team_id)
-    end
+  # Use callbacks to share common setup or constraints between actions.
+  def set_nickname
+    @nickname = Nickname.find(params[:id])
+  end
+
+  # Never trust parameters from the scary internet, only allow the white list through.
+  def nickname_params
+    params.require(:nickname).permit(:name, :team_id)
+  end
 end
