@@ -68,6 +68,12 @@ class App extends Component {
   auth = new AuthService()
 
   componentDidMount() {
+    if(this.auth.isLoggedIn()) {
+      this.fetchUserFavoriteTeams()
+    }
+  }
+
+  fetchUserFavoriteTeams = () => {
     axios.get('/user_favorite_teams',
       { headers: { Authorization: this.auth.getToken() } }
     ).then(response => {
@@ -111,7 +117,6 @@ class App extends Component {
               component={NotificationCenter}
               handleFavoriteTeamChange={this.handleFavoriteTeamChange}
               favoriteTeams={favoriteTeams}
-              />
             />
             <AdminRoute
               exact
