@@ -3,10 +3,10 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import '../styles/checkbox.scss'
 
 interface Props {
-  handleClick: () => void
-  isChecked: boolean
-  isDisabled: boolean
-  label: string
+  handleClick: () => void;
+  isChecked: boolean;
+  isDisabled: boolean;
+  label: string;
 }
 
 class Checkbox extends Component<Props> {
