@@ -1,7 +1,7 @@
 FactoryBot.define do
   factory :user_favorite_team do
     user { nil }
-    team { nil }
+    team { Team.all.sample }
     deleted { false }
   end
 end
