@@ -1,9 +1,9 @@
 FactoryBot.define do
   factory :user_notification_preference do
-    user { nil }
-    team { nil }
-    phone { 1 }
-    callback_url { "MyString" }
-    email { "MyString" }
+    user
+    team { Team.all.sample }
+    callback_url { nil }
+    email { true }
+    phone { false }
   end
 end
