@@ -1,8 +1,7 @@
 FactoryBot.define do
   factory :user_phone do
-    number { "MyString" }
-    verified { false }
+    number { "1234567890" }
+    verified_at { nil }
     verification_code { "MyString" }
-    user { nil }
   end
 end
