@@ -1,6 +1,10 @@
 FactoryBot.define do
   factory :user do
-    email { "MyString" }
-    password_digest { "MyString" }
+    user_type { UserType.regular }
+    password_digest { 'myPass' }
+    
+    after(:create) do |user, _evaluator|
+      create(:user_email, user: user)
+    end
   end
 end
